+++ opnsense_cli/commands/firewall/__init__.py
@@ -1,0 +1,13 @@
+import click
+
+
+@click.group()
+def firewall(**kwargs):
+    """
+    Execute firewall operations
+    """
+    pass
+    #ctx.obj = FirewallAliasesUtil(api_client)
+
+
+